+++ actionpack/test/controller/integration_test.rb
@@ -239,15 +239,6 @@ class SessionTest < ActiveSupport::TestCase
     end
   end
 
-  def test_deprecated_args_xml_http_request_get
-    path = "/index"; params = "blah"; headers = { location: 'blah' }
-    assert_called_with @session, :process, [:get, path, params: params, headers: headers, xhr: true] do
-      assert_deprecated(/xml_http_request/) {
-        @session.xml_http_request(:get, path, params, headers)
-      }
-    end
-  end
-
   def test_xml_http_request_post
     path = "/index"; params = "blah"; headers = { location: 'blah' }
     assert_called_with @session, :process, [:post, path, params: params, headers: headers, xhr: true] do
@@ -259,13 +250,6 @@ class SessionTest < ActiveSupport::TestCase
     path = "/index"; params = "blah"; headers = { location: 'blah' }
     assert_called_with @session, :process, [:post, path, params: params, headers: headers, xhr: true] do
       @session.post(path, params: params, headers: headers, xhr: true)
-    end
-  end
-
-  def test_deprecated_args_xml_http_request_post
-    path = "/index"; params = "blah"; headers = { location: 'blah' }
-    assert_called_with @session, :process, [:post, path, params: params, headers: headers, xhr: true] do
-      assert_deprecated(/xml_http_request/) { @session.xml_http_request(:post,path,params,headers) }
     end
   end
 
@@ -283,13 +267,6 @@ class SessionTest < ActiveSupport::TestCase
     end
   end
 
-  def test_deprecated_args_xml_http_request_patch
-    path = "/index"; params = "blah"; headers = { location: 'blah' }
-    assert_called_with @session, :process, [:patch, path, params: params, headers: headers, xhr: true] do
-      assert_deprecated(/xml_http_request/) { @session.xml_http_request(:patch,path,params,headers) }
-    end
-  end
-
   def test_xml_http_request_put
     path = "/index"; params = "blah"; headers = { location: 'blah' }
     assert_called_with @session, :process, [:put, path, params: params, headers: headers, xhr: true] do
@@ -304,13 +281,6 @@ class SessionTest < ActiveSupport::TestCase
     end
   end
 
-  def test_deprecated_args_xml_http_request_put
-    path = "/index"; params = "blah"; headers = { location: 'blah' }
-    assert_called_with @session, :process, [:put, path, params: params, headers: headers, xhr: true] do
-      assert_deprecated(/xml_http_request/) { @session.xml_http_request(:put, path, params, headers) }
-    end
-  end
-
   def test_xml_http_request_delete
     path = "/index"; params = "blah"; headers = { location: 'blah' }
     assert_called_with @session, :process, [:delete, path, params: params, headers: headers, xhr: true] do
@@ -318,38 +288,10 @@ class SessionTest < ActiveSupport::TestCase
     end
   end
 
-  def test_deprecated_xml_http_request_delete
-    path = "/index"; params = "blah"; headers = { location: 'blah' }
-    assert_called_with @session, :process, [:delete, path, params: params, headers: headers, xhr: true] do
-      assert_deprecated { @session.xml_http_request(:delete, path, params: params, headers: headers) }
-    end
-  end
-
-  def test_deprecated_args_xml_http_request_delete
-    path = "/index"; params = "blah"; headers = { location: 'blah' }
-    assert_called_with @session, :process, [:delete, path, params: params, headers: headers, xhr: true] do
-      assert_deprecated(/xml_http_request/) { @session.xml_http_request(:delete, path, params, headers) }
-    end
-  end
-
   def test_xml_http_request_head
     path = "/index"; params = "blah"; headers = { location: 'blah' }
     assert_called_with @session, :process, [:head, path, params: params, headers: headers, xhr: true] do
       @session.head(path, params: params, headers: headers, xhr: true)
-    end
-  end
-
-  def test_deprecated_xml_http_request_head
-    path = "/index"; params = "blah"; headers = { location: 'blah' }
-    assert_called_with @session, :process, [:head, path, params: params, headers: headers, xhr: true] do
-      assert_deprecated(/xml_http_request/) { @session.xml_http_request(:head, path, params: params, headers: headers) }
-    end
-  end
-
-  def test_deprecated_args_xml_http_request_head
-    path = "/index"; params = "blah"; headers = { location: 'blah' }
-    assert_called_with @session, :process, [:head, path, params: params, headers: headers, xhr: true] do
-      assert_deprecated { @session.xml_http_request(:head, path, params, headers) }
     end
   end
 end
@@ -567,18 +509,6 @@ class IntegrationProcessTest < ActionDispatch::IntegrationTest
   def test_xml_http_request_get
     with_test_route_set do
       get '/get', xhr: true
-      assert_equal 200, status
-      assert_equal "OK", status_message
-      assert_response 200
-      assert_response :success
-      assert_response :ok
-      assert_equal "JS OK", response.body
-    end
-  end
-
-  def test_deprecated_xml_http_request_get
-    with_test_route_set do
-      assert_deprecated { xhr :get, '/get' }
       assert_equal 200, status
       assert_equal "OK", status_message
       assert_response 200